fixed warnings, add flags CMakeLists.txt

+++ include/eigenData.h
@@ -28,8 +28,8 @@ namespace HydraulicNetwork {
         // std::ofstream file(fileName);
         if (!file) return;
         {
-            for (size_t i{0}; i < matrix.rows(); ++i) {
-                for (size_t j{0}; j < matrix.cols(); ++j) {
+            for (int i{0}; i < matrix.rows(); ++i) {
+                for (int j{0}; j < matrix.cols(); ++j) {
                     if (matrix(i, j) == -0) {
                         matrix(i, j) = 0;
                     }
